+++ jormungandr-lib/src/interfaces/block0_configuration/number_of_slots_per_epoch.rs
@@ -18,7 +18,7 @@ impl NumberOfSlotsPerEpoch {
     ///
     /// assert_eq!(NumberOfSlotsPerEpoch::MINIMUM, NumberOfSlotsPerEpoch::new(1).unwrap())
     /// ```
-    pub const MINIMUM: Self = NumberOfSlotsPerEpoch(MINIMUM_NUMBER_OF_SLOTS_PER_EPOCH);
+    pub const MINIMUM: Self = Self(MINIMUM_NUMBER_OF_SLOTS_PER_EPOCH);
 
     /// maximal value for the number of slots per epoch
     ///
@@ -27,7 +27,7 @@ impl NumberOfSlotsPerEpoch {
     ///
     /// assert_eq!(NumberOfSlotsPerEpoch::MAXIMUM, NumberOfSlotsPerEpoch::new(1_000_000).unwrap())
     /// ```
-    pub const MAXIMUM: Self = NumberOfSlotsPerEpoch(MAXIMUM_NUMBER_OF_SLOTS_PER_EPOCH);
+    pub const MAXIMUM: Self = Self(MAXIMUM_NUMBER_OF_SLOTS_PER_EPOCH);
 
     /// create a new `NumberOfSlotsPerEpoch` value
     ///
@@ -37,7 +37,7 @@ impl NumberOfSlotsPerEpoch {
         if v < MINIMUM_NUMBER_OF_SLOTS_PER_EPOCH || MAXIMUM_NUMBER_OF_SLOTS_PER_EPOCH < v {
             None
         } else {
-            Some(NumberOfSlotsPerEpoch(v))
+            Some(Self(v))
         }
     }
 }